key_resultsテーブルのtarget_value,actual_valueカラムの型をfloatに変更

+++ db/migrate/20170628124838_create_key_results.rb
@@ -6,8 +6,8 @@ class CreateKeyResults < ActiveRecord::Migration[5.1]
       t.integer :owner_id,      null: false
       t.integer :okr_period_id, null: false
       t.integer :progress_rate # 進捗率
-      t.integer :target_value
-      t.integer :actual_value
+      t.float :target_value
+      t.float :actual_value
       t.string  :value_unit
       t.date :expired_date
